requirements.txt
antlr4-python3-runtime==4.12.0
antlr4-tools==0.2
asgiref==3.7.2
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
black==24.3.0
bs4==0.0.1
certifi==2023.7.22
click==8.1.3
coverage==7.2.6
Django==4.2.15
exceptiongroup==1.1.1
h11==0.14.0
idna==3.7
iniconfig==2.0.0
install-jdk==1.0.2
joblib==1.2.0
lxml>=5.2.0
mccabe==0.7.0
mypy-extensions==1.0.0
nltk==3.9
outcome==1.2.0
packaging==23.1
pathspec==0.11.1
Pillow==10.3.0
platformdirs==3.2.0
pluggy==1.0.0
pycodestyle==2.10.0
pydocstyle==6.3.0
pyflakes==3.0.1
pylama==8.4.1
PySocks==1.7.1
pytest==7.3.1
pytest-cov==4.1.0
regex==2022.10.31
selenium==4.9.1
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.5.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==2.2.2
watchdog==3.0.0
word2number==1.1
wsproto==1.2.0
html-sanitizer==2.4.2
xlsxwriter==3.2.0
requests==2.32.3